Shifted stuff to BackgroundAsteroidManager
