+++ engine/src/main/java/org/destinationsol/menu/MenuObject.java
@@ -17,5 +17,6 @@ package org.destinationsol.menu;
 
 public interface MenuObject {
 
+    void update();
 
 }
