+++ engine/src/main/resources/assets/menuShips/menuships.json
@@ -1,0 +1,122 @@
+{
+  "rigidBodies": [
+    {
+      "name": "engine:imperialTiny",
+      "origin": {
+        "x": 0.6022728681564331,
+        "y": 0.5094696879386902
+      },
+      "polygons": [
+        [
+          {
+            "x": 0.6850000023841858,
+            "y": 0.32375001907348633
+          },
+          {
+            "x": 0.9399999976158142,
+            "y": 0.5012500286102295
+          },
+          {
+            "x": 0.6825000643730164,
+            "y": 0.6762500405311584
+          },
+          {
+            "x": 0.14750002324581146,
+            "y": 0.6787499785423279
+          },
+          {
+            "x": 0.15000002086162567,
+            "y": 0.3187500536441803
+          }
+        ]
+      ],
+      "circles": [],
+      "shapes": [
+        {
+          "type": "POLYGON",
+          "vertices": [
+            {
+              "x": 0.15000002086162567,
+              "y": 0.3187500536441803
+            },
+            {
+              "x": 0.14750002324581146,
+              "y": 0.6787499785423279
+            },
+            {
+              "x": 0.6825000643730164,
+              "y": 0.6762500405311584
+            },
+            {
+              "x": 0.9399999976158142,
+              "y": 0.5012500286102295
+            },
+            {
+              "x": 0.6850000023841858,
+              "y": 0.32375001907348633
+            }
+          ]
+        }
+      ]
+    },
+    {
+      "name": "engine:imperialSmall",
+      "origin": {
+        "x": 0.6022728681564331,
+        "y": 0.5094696879386902
+      },
+      "polygons": [
+        [
+          {
+            "x": 0.6850000023841858,
+            "y": 0.32375001907348633
+          },
+          {
+            "x": 0.9399999976158142,
+            "y": 0.5012500286102295
+          },
+          {
+            "x": 0.6825000643730164,
+            "y": 0.6762500405311584
+          },
+          {
+            "x": 0.14750002324581146,
+            "y": 0.6787499785423279
+          },
+          {
+            "x": 0.15000002086162567,
+            "y": 0.3187500536441803
+          }
+        ]
+      ],
+      "circles": [],
+      "shapes": [
+        {
+          "type": "POLYGON",
+          "vertices": [
+            {
+              "x": 0.15000002086162567,
+              "y": 0.3187500536441803
+            },
+            {
+              "x": 0.14750002324581146,
+              "y": 0.6787499785423279
+            },
+            {
+              "x": 0.6825000643730164,
+              "y": 0.6762500405311584
+            },
+            {
+              "x": 0.9399999976158142,
+              "y": 0.5012500286102295
+            },
+            {
+              "x": 0.6850000023841858,
+              "y": 0.32375001907348633
+            }
+          ]
+        }
+      ]
+    }
+  ]
+}
